requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.4
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.16
django-environ==0.11.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==5.0.7
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
